Update Branch_and_bound_solution.py
The branch_and_bound_solutions.py has been worked on. The algorithm now produces a feasible result. The problem is, that it cannot decrease the compensation. A steady increase of a compensation in order to include only the next OD did not work because it does not in certain circumstances find a profitable way to increase the solution. We then proceded to loop through all customers and ODs. The result was also a suboptimal solution. However tomorrow we can implement another way that also incorporates decreasing compensations. I think this might lead to an optimal solution in all cases but we will see.