Changed Gurobi_optimal_solution and branch_and_bound
1. The model in Gurobi_optimal_solution was flawed so i had to change it in order to accurately describe the problem. For that the z variable was removed and the associated constraint too. The replacements are two new constraints. In total the model now runs faster but still pretty slow for high O values.

Branch_and_bound: I observed that the optimal solution has always a certain pattern that is independent of the arrival rates but depends on the utility matrix and the opportunity costs. I concluded that there was an straightforward way to build an optimal solution without actually runnign the solver. The solution method should run pretty fast for bigger instances as well. The programming work has been done mostly and the descriptions of the methods are there. The code has yet to be tested though and there are probably many bugs in it.