Debugged Branch_and_bound
I debugged the code from yesterday and changed it completely afterwards. I experimented a little bit with the assignment strategies.  Turns out I did not have a satisfying approach yet. One attempt might be to look at the assignment matrix, that builds the optimal assignment only looking at detours. This turns out to be the optimal assignment solution and the only assignment that uniquely assigns a set of ODs to a set of customers. Maybe we can move away from this perfect assignment iteratively to get to the best assignment for one case.