Made a small neighborhood search. (In fact now it is a bit more generalized)
Generalized the neighborhood-search. It does not generate the best solutions in the world but only 2 easy neighborhoods are included. I will actually include another form of neighborhood where a more sophisticated change in stati is allowed tomorrow. Also deleted a few older methods but it seems one of the methods was used in the greedy search (which was the previous version of the neighborhood search) and used unique assignments.